Move version info from version.py to tf_ops.py
version.py file won't be available later for this package as its a singly
contained module. If it was a package, I could put it inside the package
directory, but since its not, I have to put the version info in the module
file.  So I've hardcoded __version__ to the top of tf_ops.py and I parse this
in the setup.py.